+++ src/services/sh-game/token-models/TokenModelFactory.ts
@@ -47,6 +47,7 @@ export class TokenModelFactory {
       case "Situation":
       case "WorkstationSituation":
       case "RoomWorkSituation":
+      case "SalonSituation" as any:
         return new SituationModel(
           token as Situation,
           this._container.get(API),