+++ src/types.d.ts
@@ -1,1 +1,3 @@
 declare const __APP_VERSION__: string;
+
+declare type ArrayItemOf<T> = T extends readonly (infer U)[] ? U : never;